+++ miniCP/core/less_sum_global_constraint.h
@@ -15,7 +15,7 @@ class Less_Sum_Global_Constraint: public Sum_Global_Constraint   /// This class 
     
 	public:
 		// constructors
-		Less_Sum_Global_Constraint (vector<Variable *> & var, bool strict, int cst);			///< constructs a new constraint which involves the variable in var and  compares the sum of the values of the variables in var to the constant cst with respect to < or <= if the boolean strict is true
+		Less_Sum_Global_Constraint (vector<Variable *> & var, bool strict, long cst);			///< constructs a new constraint which involves the variable in var and  compares the sum of the values of the variables in var to the constant cst with respect to < or <= if the boolean strict is true
 		Less_Sum_Global_Constraint (Less_Sum_Global_Constraint & c);  	///< constructs a new constraint by copying the constraint c
 		
 		// basic functions
@@ -27,7 +27,7 @@ class Less_Sum_Global_Constraint: public Sum_Global_Constraint   /// This class 
     bool Is_Strict_Inequality ();                 ///< returns true if the sum must be less than the constant, false otherwise
 
   private:
-    bool Compare (int sum);                     ///< returns true if the sum satisfies the constraint, false otherwise
+    bool Compare (long sum);                     ///< returns true if the sum satisfies the constraint, false otherwise
 };
 
 
@@ -62,7 +62,7 @@ inline string Less_Sum_Global_Constraint::Get_Relation ()
 }
 
 
-inline bool Less_Sum_Global_Constraint::Compare (int sum)
+inline bool Less_Sum_Global_Constraint::Compare (long sum)
 // returns true if the sum satisfies the constraint, false otherwise
 {
   return strict_inequality ? sum < constant : sum <= constant;
