+++ miniCP/core/less_weighted_sum_global_constraint.cpp
@@ -6,7 +6,7 @@
 //--------------
 
 
-Less_Weighted_Sum_Global_Constraint::Less_Weighted_Sum_Global_Constraint (vector<Variable *> & var, vector<int> & w, bool is_strict, int cst): Weighted_Sum_Global_Constraint (var,w,cst)
+Less_Weighted_Sum_Global_Constraint::Less_Weighted_Sum_Global_Constraint (vector<Variable *> & var, vector<int> & w, bool is_strict, long cst): Weighted_Sum_Global_Constraint (var,w,cst)
 // constructs a new constraint which involves the variable in var and whose relation compares the sum of the values of the variables in var weighted by weight from w to the constant cst with respect to < or <= if the boolean is_strict is true
 {	
   strict_inequality = is_strict;
@@ -28,7 +28,7 @@ Less_Weighted_Sum_Global_Constraint::Less_Weighted_Sum_Global_Constraint (Less_W
 bool Less_Weighted_Sum_Global_Constraint::Revise (CSP * pb, unsigned int var, Support * ls, Deletion_Stack * ds)
 // returns true if the application of arc-consistency on the constraint w.r.t. the variable var deletes a value in the domain of var, false otherwise
 {
-	int sum = 0;
+	long sum = 0;
 	Domain * dy;
 
 	for (unsigned int i = 0; i < arity; i++)
@@ -91,7 +91,7 @@ bool Less_Weighted_Sum_Global_Constraint::Revise (CSP * pb, unsigned int var, Su
 void Less_Weighted_Sum_Global_Constraint::Propagate (CSP * pb, Assignment & A, Support * ls, Deletion_Stack * ds, timestamp ref)
 // applies the event-based propagator of the constraint by considering the events occurred since ref
 {
-  int sum = 0;
+  long sum = 0;
   Domain * dy;
 
   for (unsigned int i = 0; i < arity; i++)  
@@ -101,7 +101,7 @@ void Less_Weighted_Sum_Global_Constraint::Propagate (CSP * pb, Assignment & A, S
       sum += dy->Get_Real_Min() * weights[i];
     else sum += dy->Get_Real_Max() * weights[i];
   }
-
+  
   for (unsigned int i = 0 ; i < arity; i++)
   {
     Domain * dx = scope_var [i]->Get_Domain();
